backend/requirements.txt
flask
flask_restful
flask_cors
SQLAlchemy
SQLAlchemy-serializer
flask-marshmallow
#psycopg2
psycopg2-binary
python-dotenv
gunicorn
sqlalchemy-utils
requests

